pyproject.toml
[project]
name = "document-processing-pipeline"
version = "0.1.0"
description = "FastAPI service that processes PDF claims using LangGraph multi-agent extraction"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "google-genai>=1.70.0",
    "langgraph>=1.1.6",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.44.0",
]
